requirements.txt
Flask
Flask-SQLAlchemy
datetime
uuid
pyjwt
pytest
pytest-mock
pylint
flask-restx>=0.2.0
Flask-Cors
flask-caching
flask-migrate
flask-bcrypt
coverage
mysqlclient>=2.0.3
mysql-connector-python
sqlalchemy-serializer

